Feat & Docs: modification de la matrice : Int to Double, ajout des methodes necessaires pour l'utilisation, ajout de la methode pour verifier la matrice, ajout des commentaires manquants

+++ src/main/java/fr/m3acnl/game/logique/Jeu.java
@@ -1,3 +1,4 @@
+
 /**
  * Classe jeu.
  *
@@ -52,7 +53,7 @@ public class Jeu {
      * @param clef clef de la matrice dans le fichier json
      * @param mat La matrice du jeu
      */
-    public Jeu(int clef, int[][] mat) {
+    public Jeu(int clef, Double[][] mat) {
         clefFichier = clef;
         instantDebut = Instant.now();
         coupJouer = new Pile();
